SRS
The purpose of the software Requirement Specification (SRS) is to specify the functionality, performance, and interface requirements of software project. All the required features of the software project expressed by the client and agreed upon by the development team shall be expressed in this document. Additional requirement of the client shall first be agreed upon by the developed team. The formal release version of the project shall meet every requirement listed in this document.